+++ plugin-gradle/src/test/java/dev/equo/ide/gradle/GradleCatalogTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2022-2023 EquoTech, Inc. and others.
+ * Copyright (c) 2022-2025 EquoTech, Inc. and others.
  *
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
@@ -13,10 +13,13 @@
  *******************************************************************************/
 package dev.equo.ide.gradle;
 
+import static org.junit.jupiter.api.condition.JRE.JAVA_17;
+
 import au.com.origin.snapshots.Expect;
 import au.com.origin.snapshots.junit5.SnapshotExtension;
 import java.io.IOException;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.EnabledForJreRange;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 @ExtendWith({SnapshotExtension.class})
@@ -34,6 +37,7 @@ public class GradleCatalogTest extends GradleHarness {
 	}
 
 	@Test
+	@EnabledForJreRange(min = JAVA_17)
 	public void simple(Expect expect) throws IOException {
 		test("jdt('4.27')", expect.scenario("jdt"));
 		test("platform('4.27')\ngradleBuildship()", expect.scenario("gradleBuildship"));