+++ solstice/src/main/java/dev/equo/ide/Catalog.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2023 EquoTech, Inc. and others.
+ * Copyright (c) 2023-2025 EquoTech, Inc. and others.
  *
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
@@ -72,8 +72,8 @@ public class Catalog implements Comparable<Catalog> {
 	public static final Catalog GRADLE_BUILDSHIP =
 			new Catalog(
 					"gradleBuildship",
-					"https://download.eclipse.org/buildship/updates/e427/releases/3.x/" + V,
-					jre11("3.1.7.v20230428-1350"),
+					"https://download.eclipse.org/buildship/updates/e431/releases/3.x/" + V,
+					jre17("3.1.10.v20240802-1211"),
 					List.of("org.eclipse.buildship.feature.group"),
 					JDT);
 
@@ -89,7 +89,7 @@ public class Catalog implements Comparable<Catalog> {
 			new Catalog(
 					"m2e",
 					"https://download.eclipse.org/technology/m2e/releases/" + V,
-					jre11("1.20.1").jre(17, "2.3.0"),
+					jre11("1.20.1").jre(17, "2.6.2"),
 					List.of("org.eclipse.m2e.feature.feature.group"),
 					JDT,
 					WST,